Relocate and rename the exceptions modules
